Move config generation code into config.go
