+++ main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"embed"
 	"encoding/json"
-	"flag"
 	"fmt"
 	"html/template"
 	"io"
@@ -16,8 +15,6 @@ import (
 	"path/filepath"
 	"strings"
 	"time"
-
-	"github.com/BurntSushi/toml"
 )
 
 type Config struct {
@@ -27,13 +24,6 @@ type Config struct {
 }
 
 var config Config
-
-const usage = `Usage:
-  -c, --config         Path to a configuration file (default: config.toml)
-  -p, --port           Port to run the server on (default: 3000)
-  -s, --serve-path     Path to serve images from (default: /i/)
-  -u, --upload-path    Path to store uploaded images (default: ./uploads/)
-`
 
 //go:embed templates
 var templatesFolder embed.FS
@@ -45,50 +35,7 @@ func init() {
 
 func main() {
 
-	// Parse the command-line flags and load the config
-	var configFile string
-	var portOpt string
-	var servePathOpt string
-	var uploadPathOpt string
-
-	flag.StringVar(&configFile, "c", "config.toml", "Path to the configuration file")
-	flag.StringVar(&configFile, "config", "config.toml", "Path to the configuration file")
-	flag.StringVar(&portOpt, "p", "3000", "Port to run the server on")
-	flag.StringVar(&portOpt, "port", "3000", "Port to run the server on")
-	flag.StringVar(&servePathOpt, "s", "", "Path to serve images from")
-	flag.StringVar(&servePathOpt, "serve-path", "", "Path to serve images from")
-	flag.StringVar(&uploadPathOpt, "u", "", "Path to store uploaded images")
-	flag.StringVar(&uploadPathOpt, "upload-path", "", "Path to store uploaded images")
-
-	flag.Usage = func() {
-		fmt.Println(usage)
-	}
-
-	if !flag.Parsed() {
-		flag.Parse()
-	}
-
-	// Load the config file if it exists otherwise use default values
-	if _, err := os.Stat(configFile); err != nil {
-		config.Port = "3000"
-		config.ServePath = "/i/"
-		config.UploadPath = "./uploads/"
-	} else {
-		config = loadConfig(configFile)
-	}
-
-	// Override the config values with the command-line flags
-	options := map[*string]*string{
-		&portOpt:       &config.Port,
-		&servePathOpt:  &config.ServePath,
-		&uploadPathOpt: &config.UploadPath,
-	}
-
-	for option, configField := range options {
-		if *option != "" {
-			*configField = *option
-		}
-	}
+	config = GenerateConfig()
 
 	// Create the upload directory if it doesn't exist
 	if _, err := os.Stat(config.UploadPath); os.IsNotExist(err) {
@@ -116,18 +63,6 @@ func main() {
 		"Upload path is %s\n",
 		config.Port, config.ServePath, config.UploadPath)
 	log.Fatal(http.ListenAndServe(":"+config.Port, nil))
-}
-
-func loadConfig(configFile string) Config {
-
-	var config Config
-
-	if _, err := toml.DecodeFile(configFile, &config); err != nil {
-		log.Fatalf("Error in parsing config file: %v", err)
-		os.Exit(1)
-	}
-
-	return config
 }
 
 func getContentType(filename string) string {
